Add linter rules
